(statscreen) make header
