+++ include/functions.h
@@ -4711,99 +4711,7 @@ void sub_8086B2C(int arg0, int arg1, u16* tilemap, int arg3, int arg4, int arg5,
 // ??? sub_8086C50(???);
 // ??? sub_8086CE8(???);
 // ??? EndBG3Slider(???);
-// ??? GetSomeUnitId(???);
-// ??? sub_8086DD8(???);
-// ??? sub_8086DE4(???);
-// ??? sub_8086DF0(???);
-// ??? sub_8086E00(???);
-// ??? sub_8086E44(???);
-// ??? sub_8086FAC(???);
-// ??? DrawStatScreenBar(???);
-// ??? DrawUnitStatScreen(???);
-// ??? DrawUnitItemScreen(???);
-// ??? DrawUnitScreenSupportList(???);
-// ??? DrawUnitWeaponRank(???);
-// ??? DrawUnitWeaponScreen(???);
-// ??? sub_80878CC(???);
-// ??? sub_8087920(???);
-// ??? sub_80879DC(???);
-// ??? sub_8087ACC(???);
-// ??? sub_8087AD8(???);
-// ??? sub_8087B40(???);
-// ??? sub_8087BA0(???);
-// ??? sub_8087BF0(???);
-// ??? sub_8087C04(???);
-// ??? sub_8087C34(???);
-// ??? sub_8087CC0(???);
-// ??? sub_8087D24(???);
-// ??? sub_8087D98(???);
-// ??? sub_8087DF8(???);
-// ??? sub_8087E28(???);
-// ??? sub_8087E7C(???);
-// ??? sub_8087EB8(???);
-// ??? sub_8087F48(???);
-// ??? sub_8087FE0(???);
-// ??? sub_8088014(???);
-// ??? sub_80880DC(???);
-// ??? sub_80881AC(???);
-// ??? sub_80881C4(???);
-// ??? sub_80881FC(???);
-// ??? sub_80882E4(???);
-// ??? sub_8088354(???);
-// ??? sub_8088384(???);
-// ??? sub_808844C(???);
-// ??? sub_80884B0(???);
-// ??? sub_8088670(???);
-// ??? sub_808873C(???);
-// ??? sub_80888B4(???);
-// ??? sub_808890C(???);
-// ??? sub_8088920(???);
-// ??? sub_808894C(???);
-// ??? MakeStatScreenRText6C(???);
-// ??? sub_8088A00(???);
-// ??? sub_8088A2C(???);
-// ??? sub_8088B08(???);
-// ??? sub_8088B40(???);
-// ??? sub_8088B94(???);
-// ??? sub_8088BD4(???);
-// ??? sub_8088C00(???);
-// ??? sub_8088C14(???);
-// ??? sub_8088C48(???);
-// ??? sub_8088CFC(???);
-// ??? sub_8088D3C(???);
-// ??? sub_8088D64(???);
-// ??? sub_8088DB8(???);
-void sub_8088DE0(int x, int y, int msgid);
-// ??? sub_8088E14(???);
-void sub_8088E60(int x, int y, int item);
-// ??? sub_8088E9C(???);
-// ??? sub_8088F68(???);
-void sub_8089018(void);
-// ??? sub_808903C(???);
-// ??? sub_8089060(???);
-// ??? sub_8089088(???);
-// ??? sub_8089150(???);
-// void Create6CRText(const struct HelpBoxInfo* helpinfo, struct Proc* parent); // StartMoveableHelpBox
-// ??? sub_8089188(???);
-// ??? sub_80891AC(???);
-// ??? sub_8089210(???);
-// ??? sub_808929C(???);
-// ??? sub_80892C0(???);
-// ??? sub_80892D0(???);
-// ??? sub_8089320(???);
-// ??? sub_8089354(???);
-// ??? sub_8089384(???);
-// ??? sub_80893B4(???);
-// ??? sub_80893E4(???);
-// ??? sub_8089430(???);
-// ??? sub_8089454(???);
-// ??? Loop6C_8A00B20_UpdateOAMData(???);
-// ??? sub_80894AC(???);
-// ??? sub_80894E0(???);
-// ??? sub_808953C(???);
-// ??? sub_8089570(???);
-// ??? sub_8089588(???);
-// ??? sub_80895A8(???);
+
 // ??? sub_80895B4(???);
 // ??? sub_8089624(???);
 // ??? sub_808966C(???);
@@ -5619,7 +5527,7 @@ void sub_80A4AA4(void);
 void DeclareCompletedPlaythrough(void);
 // ??? sub_80A4CB4(???);
 // ??? sub_80A4CD8(???);
-// ??? BWL_GetEntry(???);
+struct UnitUsageStats* BWL_GetEntry(int pid);
 // ??? sub_80A4D28(???);
 // ??? Set0203EDB4(???);
 // ??? Save0203EDB4(???);
@@ -7309,5 +7217,6 @@ void SoundBiasSet();
 void ClearOAMBuffer();
 void TileMap_FillRect(u16 *dest, int width, int height, int fillValue);
 void TileMap_CopyRect(u16* src, u16* dst, int width, int height);
+void CallARM_FillTileRect(u16* tilemap, const void* tsa, int tileref);
 
 #endif  // GUARD_FUNCTIONS_H
