+++ include/types.h
@@ -459,6 +459,25 @@ struct MapChange
     /* 08 */ const void* data;
 };
 
+struct UnitUsageStats
+{
+	/* 000 */ unsigned lossAmt     : 8;
+	/* 008 */ unsigned favval      : 16;
+	/* 024 */ unsigned actAmt      : 8;
+	/* 032 */ unsigned statViewAmt : 8;
+	/* 040 */ unsigned deathLoc    : 6;
+	/* 046 */ unsigned deathTurn   : 10;
+	/* 056 */ unsigned deployAmt   : 6;
+	/* 062 */ unsigned moveAmt     : 10;
+	/* 072 */ unsigned deathCause  : 4;
+	/* 076 */ unsigned expGained   : 12;
+	/* 088 */ unsigned winAmt      : 10;
+	/* 098 */ unsigned battleAmt   : 12;
+	/* 110 */ unsigned killerPid   : 9;
+	/* 119 */ unsigned deathSkirm  : 1;
+	/* 120 */ /* 8bit pad */
+};
+
 enum { UNIT_SUPPORT_MAX_COUNT = 7 };
 
 enum
